Move main function from Application to RunGame
