+++ src/main/java/is/ru/tictactoe/RunGame.java
@@ -1,0 +1,12 @@
+package is.ru.tictactoe;
+
+public class RunGame {
+
+    public static void main (String[] args) {
+        Game game = new Game();
+        UI ui = new UI();
+        Application app = new Application(game,ui);
+
+        app.runApp();
+    }
+}
